# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: RentFlow API
  description: |
    Multi-tenant Rent Management SaaS API.

    **Rules enforced at every endpoint:**
    - `public_id` is used in all URLs and responses — internal `id` is never exposed
    - All resources are scoped to the authenticated owner via JWT
    - Soft-deleted resources (`is_active=false`) are treated as not found (404)
    - All monetary values are numbers (`NUMERIC(10,2)` precision)
  version: "1.0.0"
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # --------------------------------------------------------------
  # Reusable path parameters
  # --------------------------------------------------------------
  parameters:
    page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    building_public_id:
      name: building_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    apartment_public_id:
      name: apartment_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    tenant_public_id:
      name: tenant_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Authentication required"
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Access denied"
    NotFound:
      description: Resource not found or soft-deleted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Not found"
    BadRequest:
      description: Business logic error (e.g. duplicate unit, conflicting dates, default category delete)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'

  # --------------------------------------------------------------
  # Schemas
  # --------------------------------------------------------------
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
          example: null
        message:
          type: string
    ValidationErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        message:
          type: string
          example: "Validation failed"
    Pagination:
      type: object
      required: [total, page, page_size]
      properties:
        total:
          type: integer
          description: Total number of matching records
        page:
          type: integer
        page_size:
          type: integer

    # ── Auth ───────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [full_name, email, password]
      properties:
        full_name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    OwnerData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    TokenData:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        owner:
          $ref: '#/components/schemas/OwnerData'

    # ── Buildings ──────────────────────────────────────────────
...
    ApartmentCreate:
      type: object
      required: [unit_number, floor]
      properties:
        unit_number:
          type: string
          maxLength: 50
        floor:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [vacant, occupied]
          default: vacant
    ApartmentUpdate:
      type: object
      properties:
        unit_number:
          type: string
          maxLength: 50
        floor:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [vacant, occupied]
    ApartmentData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        building_public_id:
          type: string
          format: uuid
        unit_number:
          type: string
        floor:
          type: integer
        status:
          type: string
          enum: [vacant, occupied]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Tenants ────────────────────────────────────────────────
...
    AgreementCreate:
      type: object
      required: [rent_amount, start_date]
      properties:
        rent_amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
    AgreementData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        tenant_public_id:
          type: string
          format: uuid
        rent_amount:
          type: number
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Monthly Dues ───────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new owner account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: Account created, JWT returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/TokenData'
                  message: { type: string, example: "Account created" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/TokenData'
                  message: { type: string, example: "Login successful" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  # ==============================================================
  # BUILDINGS
  # ==============================================================
...
  /buildings/{building_public_id}/apartments:
    parameters:
      - $ref: '#/components/parameters/building_public_id'
    get:
      tags: [Apartments]
      summary: List apartments in a building
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
        - name: status
          in: query
          schema:
            type: string
            enum: [vacant, occupied]
      responses:
        "200":
          description: Apartments list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ApartmentData'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Apartments]
      summary: Create an apartment in a building
      description: "unit_number must be unique within the building."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApartmentCreate'
      responses:
        "201":
          description: Apartment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/ApartmentData'
                  message: { type: string, example: "Apartment created" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /buildings/{building_public_id}/apartments/{apartment_public_id}:
    parameters:
      - $ref: '#/components/parameters/building_public_id'
      - $ref: '#/components/parameters/apartment_public_id'
    get:
      tags: [Apartments]
      summary: Get a single apartment
      responses:
        "200":
          description: Apartment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/ApartmentData'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Apartments]
      summary: Update apartment details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApartmentUpdate'
      responses:
        "200":
          description: Apartment updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/ApartmentData'
                  message: { type: string, example: "Apartment updated" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Apartments]
      summary: Soft delete an apartment
      description: "Returns 400 if the apartment has an active tenant."
      responses:
        "200":
          description: Apartment deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { nullable: true, example: null }
                  message: { type: string, example: "Apartment deactivated" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

  # ==============================================================
  # TENANTS
  # ==============================================================
...
  /tenants/{tenant_public_id}/agreements:
    parameters:
      - $ref: '#/components/parameters/tenant_public_id'
    get:
      tags: [Agreements]
      summary: List all agreements for a tenant
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        "200":
          description: Agreements list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/AgreementData'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Agreements]
      summary: Create a new agreement for a tenant (e.g. rent increase)
      description: |
        Creates a new agreement record. The new agreement becomes the active one.
        Monthly dues generated after this date will use the new rent_amount.
        Returns 400 if start_date overlaps with an existing active agreement.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgreementCreate'
      responses:
        "201":
          description: Agreement created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/AgreementData'
                  message: { type: string, example: "Agreement created" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  # ==============================================================
  # MONTHLY DUES
  # ==============================================================
...